requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
langchain>=0.1.0
langgraph>=0.2.0
langgraph-cli[inmem]>=0.1.50
langgraph-api>=0.7.3
langchain-openai>=0.0.5
langchain-community>=0.0.20
langsmith>=0.0.87
qdrant-client>=1.7.0
sentence-transformers>=2.2.2
sqlalchemy>=2.0.0
pandas>=2.0.0
openpyxl>=3.1.0
duckdb>=0.9.0
plotly>=5.18.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
rank-bm25>=0.2.2
httpx>=0.26.0
streamlit>=1.30.0
requests>=2.31.0
python-multipart>=0.0.6
starlette>=0.35.0
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-env>=1.1.0